further fix for if coverenddate is before cover start date
