+++ e2e-tests/portal/cypress/integration/journeys/checker/cover-end-date/cover-end-date-validation.spec.js
@@ -86,6 +86,7 @@ context('Checker tries to submit a deal that has changed/newly issued facilities
 
           const { mockFacilities } = dealAllAcknowledged;
           mockFacilities[0].status = CONSTANTS.DEALS.DEAL_STATUS.READY_FOR_APPROVAL;
+          mockFacilities[2].status = CONSTANTS.DEALS.DEAL_STATUS.READY_FOR_APPROVAL;
 
           cy.createFacilities(dealId, mockFacilities, BANK1_MAKER1).then((createdFacilities) => {
             const bonds = createdFacilities.filter((f) => f.type === CONSTANTS.FACILITY.FACILITY_TYPE.BOND);
