+++ e2e-tests/portal/cypress/integration/journeys/checker/cover-end-date/deal.js
@@ -16,12 +16,13 @@ const deal = {
       bondType: FIXTURE_CONSTANTS.BSS_FACILITY_TYPE.BID_BOND,
       facilityStage: CONSTANTS.FACILITY.FACILITY_STAGE.ISSUED,
       hasBeenIssued: true,
-      'requestedCoverStartDate-day': '',
-      'requestedCoverStartDate-month': '',
-      'requestedCoverStartDate-year': '',
+      'requestedCoverStartDate-day': '13',
+      'requestedCoverStartDate-month': '09',
+      'requestedCoverStartDate-year': '2022',
       'coverEndDate-day': '06',
       'coverEndDate-month': '01',
       'coverEndDate-year': '2021',
+      requestedCoverStartDate: '1663063419194',
       name: '1234',
       bondBeneficiary: '',
       guaranteeFeePayableByBank: '18.0000',
@@ -88,9 +89,10 @@ const deal = {
       createdDate: date.valueOf(),
       facilityStage: CONSTANTS.FACILITY.FACILITY_STAGE.UNCONDITIONAL,
       hasBeenIssued: true,
-      'requestedCoverStartDate-day': '',
-      'requestedCoverStartDate-month': '',
-      'requestedCoverStartDate-year': '',
+      'requestedCoverStartDate-day': '13',
+      'requestedCoverStartDate-month': '09',
+      'requestedCoverStartDate-year': '2022',
+      requestedCoverStartDate: '1663063419194',
       'coverEndDate-day': '06',
       'coverEndDate-month': '01',
       'coverEndDate-year': '2021',
