+++ portal-api/src/v1/validation/facility-change-cover-start-date-rules/index.js
@@ -12,7 +12,7 @@ module.exports = (
     errorList,
     deal,
   );
-  errorList = coverDatesRules(submittedValues, errorList);
+  errorList = coverDatesRules(submittedValues, deal, errorList);
 
   return errorList;
 };
