+++ portal-api/src/v1/validation/fields/cover-dates.js
@@ -4,8 +4,9 @@ const {
   dateHasAllValues,
 } = require('./date');
 const { formattedTimestamp } = require('../../facility-dates/timestamp');
+const isReadyForValidation = require('../helpers/isReadyForValidation.helper');
 
-module.exports = (submittedValues, errorList) => {
+module.exports = (submittedValues, deal, errorList) => {
   const newErrorList = errorList;
 
   const requestedCoverStartDateTimestamp = formattedTimestamp(submittedValues.requestedCoverStartDate);
@@ -16,36 +17,38 @@ module.exports = (submittedValues, errorList) => {
     'coverEndDate-year': coverEndDateYear,
   } = submittedValues;
 
-  const hasValidRequestedCoverStartDate = (requestedCoverStartDateTimestamp && !newErrorList.requestedCoverStartDate);
+  if (isReadyForValidation(deal, submittedValues)) {
+    const hasValidRequestedCoverStartDate = (requestedCoverStartDateTimestamp && !newErrorList.requestedCoverStartDate);
 
-  const hasValidCoverEndDate = dateHasAllValues(
-    coverEndDateDay,
-    coverEndDateMonth,
-    coverEndDateYear,
-  ) && !newErrorList.coverEndDate;
+    const hasValidCoverEndDate = dateHasAllValues(
+      coverEndDateDay,
+      coverEndDateMonth,
+      coverEndDateYear,
+    ) && !newErrorList.coverEndDate;
 
-  const hasValidCoverStartAndEndDates = (hasValidRequestedCoverStartDate && hasValidCoverEndDate);
+    const hasValidCoverStartAndEndDates = (hasValidRequestedCoverStartDate && hasValidCoverEndDate);
 
-  if (hasValidCoverStartAndEndDates) {
-    const requestedCoverStartDateDay = moment(requestedCoverStartDateTimestamp).format('DD');
-    const requestedCoverStartDateMonth = moment(requestedCoverStartDateTimestamp).format('MM');
-    const requestedCoverStartDateYear = moment(requestedCoverStartDateTimestamp).format('YYYY');
+    if (hasValidCoverStartAndEndDates) {
+      const requestedCoverStartDateDay = moment(requestedCoverStartDateTimestamp).format('DD');
+      const requestedCoverStartDateMonth = moment(requestedCoverStartDateTimestamp).format('MM');
+      const requestedCoverStartDateYear = moment(requestedCoverStartDateTimestamp).format('YYYY');
 
-    const requestedCoverStartDate = `${requestedCoverStartDateYear}-${requestedCoverStartDateMonth}-${requestedCoverStartDateDay}`;
-    const coverEndDate = `${coverEndDateYear}-${coverEndDateMonth}-${coverEndDateDay}`;
+      const requestedCoverStartDate = `${requestedCoverStartDateYear}-${requestedCoverStartDateMonth}-${requestedCoverStartDateDay}`;
+      const coverEndDate = `${coverEndDateYear}-${coverEndDateMonth}-${coverEndDateDay}`;
 
-    if (moment(coverEndDate).isBefore(requestedCoverStartDate)) {
-      newErrorList.coverEndDate = {
-        text: 'Cover End Date cannot be before Requested Cover Start Date',
-        order: orderNumber(newErrorList),
-      };
-    }
+      if (moment(coverEndDate).isBefore(requestedCoverStartDate)) {
+        newErrorList.coverEndDate = {
+          text: 'Cover End Date cannot be before Requested Cover Start Date',
+          order: orderNumber(newErrorList),
+        };
+      }
 
-    if (moment(coverEndDate).isSame(requestedCoverStartDate)) {
-      newErrorList.coverEndDate = {
-        text: 'Cover End Date must be after the Requested Cover Start Date',
-        order: orderNumber(newErrorList),
-      };
+      if (moment(coverEndDate).isSame(requestedCoverStartDate)) {
+        newErrorList.coverEndDate = {
+          text: 'Cover End Date must be after the Requested Cover Start Date',
+          order: orderNumber(newErrorList),
+        };
+      }
     }
   }
