+++ portal-api/src/v1/validation/fields/cover-end-date.js
@@ -4,19 +4,7 @@ const {
   dateHasAllValues,
   dateValidationText,
 } = require('./date');
-
-const { SUBMISSION_TYPE } = require('../../../constants/deal');
-const { DEAL_STATUS } = require('../../../constants/facilities');
-
-const isReadyForValidation = (deal, submittedValues) => {
-  const { submissionType } = deal;
-  const { status } = submittedValues;
-
-  if ((submissionType === SUBMISSION_TYPE.AIN || submissionType === SUBMISSION_TYPE.MIN) && status === DEAL_STATUS.ACKNOWLEDGED) {
-    return false;
-  }
-  return true;
-};
+const isReadyForValidation = require('../helpers/isReadyForValidation.helper');
 
 module.exports = (submittedValues, deal, errorList) => {
   const newErrorList = errorList;
