+++ portal-api/src/v1/validation/loan-rules/facility-stage-unconditional/index.js
@@ -16,7 +16,7 @@ module.exports = (loan, errorList, deal) => {
     newErrorList = name(loan, newErrorList);
     newErrorList = requestedCoverStartDateRules(loan, deal, newErrorList);
     newErrorList = coverEndDate(loan, deal, newErrorList);
-    newErrorList = coverDates(loan, newErrorList);
+    newErrorList = coverDates(loan, deal, newErrorList);
     newErrorList = disbursementAmount(loan, newErrorList);
   }
